% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{data_plot}
\alias{data_plot}
\title{The summary of variables by plot}
\usage{
data_plot(data, data_class)
}
\arguments{
\item{data}{a data frame}

\item{data_class}{a variable class}
}
\value{
a plot
}
\description{
This function performs a plot analysis of variables
}
\examples{
data_plot(mtcars, "double")

}
