% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tibble.R
\name{data_tibble}
\alias{data_tibble}
\title{The summary of variables}
\usage{
data_tibble(data)
}
\arguments{
\item{a}{data frame}
}
\value{
a table with data variables and name/class/number of unique values/number of NA of variables
}
\description{
The summary of variables
}
\examples{
data_tibble(starwars)
data_tibble(billboard)
}
