% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pgoment-package.R
\docType{package}
\name{hw04pgoment-package}
\alias{hw04pgoment}
\alias{hw04pgoment-package}
\title{hw04pgoment: The package summary the data frame (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Miantong Guo \email{mg4271a@american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
