% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw04pgoment-package.R
\docType{package}
\name{hw04pgoment-package}
\alias{hw04pgoment}
\alias{hw04pgoment-package}
\title{hw04pgoment: Summary the variables of a data frame}
\description{
The package can summay variables of your choosen data frame, it also can provides plots of variables.
}
\author{
\strong{Maintainer}: Miantong Guo \email{guomiantong@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
